Split Direct2D bitmap renderer from drag image creator
This splits off the Direct2D bitmap renderer code in the drag image creator into a separate class so that it can be reused.
