+++ ui_helpers.vcxproj
@@ -155,6 +155,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="container_window.h" />
+    <ClInclude Include="d2d_bitmap_renderer.h" />
     <ClInclude Include="dcomp_utils.h" />
     <ClInclude Include="dialog.h" />
     <ClInclude Include="direct_2d.h" />
@@ -197,6 +198,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="container_window.cpp" />
+    <ClCompile Include="d2d_bitmap_renderer.cpp" />
     <ClCompile Include="dcomp_utils.cpp" />
     <ClCompile Include="dialog.cpp" />
     <ClCompile Include="direct_2d.cpp" />
